+++ lib/src/element_params/sub/structure/split_params.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:simple_managers/simple_managers.dart';
 
 ///
 /// (en) A wrapper class for parameters used in classes with the same prefix name (XXXParamsWrapper of X).
@@ -19,14 +20,14 @@ class SplitParamsWrapper {
 /// (ja) 接頭名(XXXParams of X)が同じクラスで利用するパラメータを保持するためのクラス。
 ///
 class SplitParams {
+  ValueManager? manager;
   static const double defBarSize = 8;
   static const double defClampMin = 0.1;
   static const double defClampMax = 0.9;
+  static const double defSplitRatio = 0.5;
   Key? key;
   // If horizontal, children is set the horizontal layout.
   Axis axis = Axis.horizontal;
-  // The split ratio.
-  double ratio = 0.5;
   SystemMouseCursor cursor = SystemMouseCursors.grab;
   double barSize = defBarSize;
   Color color = Colors.grey;